Converted ProcessorError to ProcessorInterrupt

+++ src-tauri/src/services/processors/types/data_confirmation.rs
@@ -3,7 +3,7 @@ use serde::Serialize;
 #[derive(Debug, Clone, Serialize)]
 pub struct DataConfirmation {
     pub subdomain: Option<String>,
-    pub locations_count: Option<usize>,
+    pub locations_count: Option<u32>,
     pub days: Option<u32>,
     pub appointment_type_name: Option<String>,
 }
